// app/types.ts
export interface FormData {
    fullName: string;
    dateOfBirth: string;
    nationality: string;
    email: string;
    phone: string;
    departureDate: string;
    returnDate: string;
    accommodation: string;
    specialRequests: string;
    healthDeclaration: string;
    emergencyContact: string;
    medicalConditions: string;
  }